role

+++ ApplicationLogic/src/main/java/dk/via/sep3/group1/applicationlogic/model/CreditCard.java
@@ -72,7 +72,7 @@ public class CreditCard {
         this.securityCode = securityCode;
     }
 
-    public int getAmountOfMoney() {
+    public float getAmountOfMoney() {
         return amountOfMoney;
     }
